Move utils that are only used in one plugin into their plugins.
