+++ src/core/getTimeStamp.js
